+++ randovania/games/samus_returns/gui/dialog/cosmetic_patches_dialog.py
@@ -1,11 +1,12 @@
 from __future__ import annotations
 
 import dataclasses
+import functools
 
 from PySide6 import QtGui, QtWidgets
 
 from randovania.games.samus_returns.gui.generated.msr_cosmetic_patches_dialog_ui import Ui_MSRCosmeticPatchesDialog
-from randovania.games.samus_returns.layout.msr_cosmetic_patches import MSRCosmeticPatches, MSRRoomGuiType
+from randovania.games.samus_returns.layout.msr_cosmetic_patches import MSRCosmeticPatches, MSRRoomGuiType, MusicMode
 from randovania.gui.dialog.base_cosmetic_patches_dialog import BaseCosmeticPatchesDialog
 from randovania.gui.lib import signal_handling
 from randovania.gui.lib.signal_handling import set_combo_with_value
@@ -21,6 +22,12 @@ class MSRCosmeticPatchesDialog(BaseCosmeticPatchesDialog, Ui_MSRCosmeticPatchesD
 
         for room_gui_type in MSRRoomGuiType:
             self.room_names_dropdown.addItem(room_gui_type.long_name, room_gui_type)
+
+        self.radio_buttons = {
+            MusicMode.VANILLA: self.vanilla_music_option,
+            MusicMode.TYPE: self.type_music_option,
+            MusicMode.FULL: self.full_music_option,
+        }
 
         self.connect_signals()
         self.on_new_cosmetic_patches(current)
@@ -60,6 +67,13 @@ class MSRCosmeticPatchesDialog(BaseCosmeticPatchesDialog, Ui_MSRCosmeticPatchesD
 
         signal_handling.on_combo(self.room_names_dropdown, self._on_room_name_mode_update)
 
+        for music_mode, radio_button in self.radio_buttons.items():
+            radio_button.toggled.connect(functools.partial(self._on_music_option_changed, music_mode))
+
+    def _on_music_option_changed(self, option: MusicMode, value: bool) -> None:
+        if value:
+            self._cosmetic_patches = dataclasses.replace(self._cosmetic_patches, music=option)
+
     def on_new_cosmetic_patches(self, patches: MSRCosmeticPatches) -> None:
         self.custom_laser_color_check.setChecked(patches.use_laser_color)
         self.custom_energy_tank_color_check.setChecked(patches.use_energy_tank_color)
@@ -81,6 +95,9 @@ class MSRCosmeticPatchesDialog(BaseCosmeticPatchesDialog, Ui_MSRCosmeticPatchesD
             box.setStyleSheet(style)
 
         set_combo_with_value(self.room_names_dropdown, patches.show_room_names)
+
+        for music_mode, radio_button in self.radio_buttons.items():
+            radio_button.setChecked(music_mode == patches.music)
 
     def _open_color_picker(self, color: tuple, propertyName: str) -> None:
         picker_result = QtWidgets.QColorDialog.getColor(QtGui.QColor(*color))